# Optimized BVH Construction & Rendering Performance