--- a/hittable.h
+++ b/hittable.h
@@ -6,7 +6,7 @@ class material;
 class hit_record
 {
 public:
-    Vector3f p; // hit point
+    Vector3f p;      // hit point
     Vector3f normal; // face normal
     shared_ptr<material> mat;
     float t;
@@ -19,6 +19,8 @@ public:
     virtual ~hittable() = default;
 
     virtual bool hit(const ray &r, interval ray_t, hit_record &rec) const = 0;
+
+    virtual bbox get_bbox() const = 0;
 };
 
 #endif