--- a/hittable_list.h
+++ b/hittable_list.h
@@ -2,8 +2,9 @@
 #define HITTABLE_LIST_H
 
 #include "hittable.h"
+#include "bvh.h"
 
-class hittable_list : public hittable
+class hittable_list
 {
 public:
     std::vector<shared_ptr<hittable>> objects;
@@ -18,23 +19,9 @@ public:
         objects.push_back(object);
     }
 
-    bool hit(const ray &r, interval ray_t, hit_record &rec) const override
+    shared_ptr<BVHNode> create_bvh_tree() const
     {
-        hit_record temp_rec;
-        bool hit_anything = false;
-        auto closest_so_far = ray_t.max;
-
-        for (const auto &object : objects)
-        {
-            if (object->hit(r, interval(ray_t.min, closest_so_far), temp_rec))
-            {
-                hit_anything = true;
-                closest_so_far = temp_rec.t;
-                rec = temp_rec;
-            }
-        }
-
-        return hit_anything;
+        return std::make_shared<BVHNode>(objects, 0, objects.size());
     }
 };