--- a/hittable_list.h
+++ b/hittable_list.h
@@ -23,6 +23,26 @@ public:
     {
         return std::make_shared<BVHNode>(objects, 0, objects.size());
     }
+
+    // 保留此方法，便于对比性能差距
+    bool hit(const ray &r, interval ray_t, hit_record &rec) const
+    {
+        hit_record temp_rec;
+        bool hit_anything = false;
+        auto closest_so_far = ray_t.max;
+
+        for (const auto &object : objects)
+        {
+            if (object->hit(r, interval(ray_t.min, closest_so_far), temp_rec))
+            {
+                hit_anything = true;
+                closest_so_far = temp_rec.t;
+                rec = temp_rec;
+            }
+        }
+
+        return hit_anything;
+    }
 };
 
 #endif