--- a/sphere.h
+++ b/sphere.h
@@ -9,6 +9,11 @@ public:
     sphere(const Vector3f &center, float radius, shared_ptr<material> mat)
         : center(center), radius(std::fmax(0, radius)), mat(mat) {}
 
+    bbox get_bbox() const override
+    {
+        return bbox(center.array() - radius, center.array() + radius);
+    }
+
     bool hit(const ray &r, interval ray_t, hit_record &rec) const override
     {
         Vector3f oc = center - r.origin();