… .gitignore
… CMakeLists.txt
… doc/1.png
… readme.md
… timer.h
… triangle.h