Issue #2 

Implemented **BVH with SAH (Surface Area Heuristic)** and enabled compiler optimizations (`-O3 -march=native`), significantly improving ray tracing performance.  

### **Key Improvements**  

1. **BVH Construction Optimizations**  
   - Implemented **SAH splitting strategy** to reduce redundant ray-AABB intersection tests  
   - Dynamically selects optimal split axis/position during recursive BVH builds  

2. **Compiler Optimizations**  
   - Enabled `-O3` and `-march=native`, improving **Eigen vector ops by 5-10x**  
   - Added `EIGEN_NO_DEBUG` to disable runtime checks  

3. **Performance Gains**  
   | Before (Debug) | After (Release + SAH) |  
   |----------------|-----------------------|  
   | ~120ms/frame   | ~12ms/frame (**10x faster**) |  

### **Validation**  
- **Test Scene**: 100k-triangle complex model  
- **Results**:  
  - BVH build time ↓ **30%** (SAH optimization)  
  - Render time/frame ↓ **90%** (Eigen + compiler opts)  

### **Future Work**  
- Incremental BVH updates for dynamic scenes  
- Manual SIMD optimizations (AVX/NEON intrinsics)  

**Review Focus**:  
- Verify `BVHNode` split logic correctness  
- Check SAH cost calculation robustness  

**Impact**: Enables interactive rendering for moderately complex scenes.